VideoIngestion: Exiting from the SIGTERM signal handler (#250)
* VideoIngestion: Exiting from the SIGTERM signal handler

It was found that on a long run, visualhmi client exited with
"Subscription Id deleted" log and ia_video_ingestion was stuck
at "ETA killed.." log. So added in sys.exit() to exit of the
container from the SIGTERM signal handler. The SIGTERM might have
originated from the docker or kernel as there was no user
interference

Signed-off-by: Vinod K B <vinod.k.b@intel.com>