VideoIngestion: Remove the -pg flag in WITH_PROFILE
The switch WITH_PROFILE was enabled for the benchmarking scenario and
this -pg flag was causing restart of containers in IPC mode publish.
Commenting off the -pg flag addition for avoiding this issue. A better
fix for this is to have a different macro for benchmarking fps
calculations. This is to be taken up in the 2.2 release with full
validation.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>