VideoIngestion: Readme change to provide usage of Gstreamer's queue element
With the addition of queue element one can control the max size of the
buffers stored in the queue and choose to either drop the upstreaming
or downstreaming buffers. With GVA use case one may notice VideoIngestion
stopping to publish frames as the inferencing might take more time due
to which the buffers can get queued up and stop the ingestion of frames.
In such scenarios this element can be used/

Signed-off-by: sandeshk <sandesh.kumar.s@intel.com>