VideoIngestion: Handle image format gracefully in gstreamer ingestor
The only supported image format in gstreamer ingestor is BGR. Any other
format will not be supported and will be logged as an error. This check
will be added only for the first frame which is received in the gstreamer
ingestor as it can be a overhead if we validate the format for each frame.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>