video-ingestion: Replace worker safety video file
Worker safety video file (MP4) has been replaced with
the transcoded version (AVI). The reason is MP4 video
files generate a EOS signal thereby termination the
gstreamer pipeline due to which multifilesrc element
is not able to loop the video. Readme has been updated
with the know issue of looping with MP4 videos.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>