video-ingestion: VI readme updates
Corrected the info regarding inter packet delay
(interpacketdealy -> packet-delay).

Avoided hardcording pixel-format to `ycbcr422_8`
so that the user verifies the supported pixel formats
of the camera and inputs in the gstreamer pipeline

Removed readme line which states hw trigger timeout
value will only function in GST INFO level for gencamsrc
as a fix was added for that.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>